+++ src/main/java/org/elasticsearch/script/javascript/JavascriptNashornScriptEngineService.java
@@ -19,12 +19,8 @@
 
 package org.elasticsearch.script.javascript;
 
-import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
-import java.util.ArrayList;
-import java.util.Collection;
 import java.util.HashMap;
-import java.util.Iterator;
 import java.util.Map;
 
 import javax.script.Bindings;
@@ -33,142 +29,113 @@ import javax.script.CompiledScript;
 import javax.script.ScriptContext;
 import javax.script.ScriptEngine;
 import javax.script.ScriptException;
+import javax.script.SimpleScriptContext;
 
-import org.apache.lucene.index.AtomicReaderContext;
-import org.apache.lucene.search.Scorer;
 import org.elasticsearch.common.Nullable;
 import org.elasticsearch.common.component.AbstractComponent;
 import org.elasticsearch.common.inject.Inject;
 import org.elasticsearch.common.settings.Settings;
-import org.elasticsearch.plugin.javascript.JavascriptPlugin;
 import org.elasticsearch.script.ExecutableScript;
 import org.elasticsearch.script.ScriptEngineService;
 import org.elasticsearch.script.SearchScript;
-import org.elasticsearch.script.javascript.support.SilentScriptContext;
+import org.elasticsearch.script.javascript.support.NashornExecutableScript;
+import org.elasticsearch.script.javascript.support.NashornScriptValueConverter;
+import org.elasticsearch.script.javascript.support.NashornSearchScript;
 import org.elasticsearch.search.lookup.SearchLookup;
 
+/**
+ * Nashorn ScriptEngineService.
+ */
 public class JavascriptNashornScriptEngineService extends AbstractComponent
         implements ScriptEngineService {
 
-    public static class NashornScriptValueConverter {
-        private static Class<?> CLASS_SCRIPT_OBJECT;
-        private static Method METHOD_SCRIPT_OBJECT_IS_ARRAY;
-        private static Method METHOD_SCRIPT_OBJECT_ENTRY_SET;
-        private static Method METHOD_SCRIPT_OBJECT_SIZE;
+	/**
+	 * Global objects are not thread-safe.
+	 * We create one ScriptEngine per thread and configure the ScriptEngine to reuse the same Global
+	 * with --global-per-engine
+	 */
+    private static final ThreadLocal<ScriptEngine> currentEngine = new ThreadLocal<ScriptEngine>();
+    /**
+     * CompiledScripts actually are referring to their Global object.
+     * If one attempts to execute a compiled Script in a different thread it gets recompiled.
+     * So we cache the script compiled object for each thread.
+     */
+    private static final ThreadLocal<Map<String,CompiledScript>> compiledScripts = new ThreadLocal<Map<String,CompiledScript>>();
 
-        private static Class<?> CLASS_ARRAY_LIKE_ITERATOR;
-        private static Method METHOD_ARRAY_ITERATOR;
-
-        private NashornScriptValueConverter() {
-        }
-        static void setup() {
-            try {
-                // Use introspection to avoid accessing nashorn specific classes.
-                // This way we can compile with jdk6
-                CLASS_SCRIPT_OBJECT = JavascriptNashornScriptEngineService.class.getClassLoader().loadClass("jdk.nashorn.internal.runtime.ScriptObject");
-                METHOD_SCRIPT_OBJECT_IS_ARRAY = CLASS_SCRIPT_OBJECT.getMethod("isArray");
-                METHOD_SCRIPT_OBJECT_ENTRY_SET = CLASS_SCRIPT_OBJECT.getMethod("entrySet");
-                METHOD_SCRIPT_OBJECT_SIZE = CLASS_SCRIPT_OBJECT.getMethod("size");
-
-                CLASS_ARRAY_LIKE_ITERATOR = JavascriptNashornScriptEngineService.class.getClassLoader().loadClass("jdk.nashorn.internal.runtime.arrays.ArrayLikeIterator");
-                METHOD_ARRAY_ITERATOR = CLASS_ARRAY_LIKE_ITERATOR.getDeclaredMethod("arrayLikeIterator", Object.class);
-            } catch (Throwable t) {
-                throw new RuntimeException(t);
-            }
-        }
-        public static Object unwrapValue(Object value) {
-            if (value == null) {
-                return null;
-            } else if (value instanceof Object[]) {
-                Object[] array = (Object[]) value;
-                ArrayList<Object> list = new ArrayList<Object>(array.length);
-                for (int i = 0; i < array.length; i++) {
-                    list.add(unwrapValue(array[i]));
-                }
-                value = list;
-            } else if (CLASS_SCRIPT_OBJECT.isInstance(value)) {
-                try {
-                    if ((Boolean)METHOD_SCRIPT_OBJECT_IS_ARRAY.invoke(value)) {
-                        ArrayList<Object> list = new ArrayList<Object>();
-                        Iterator<Object> it = (Iterator<Object>) METHOD_ARRAY_ITERATOR.invoke(null, value);
-                        while (it.hasNext()) {
-                            list.add(unwrapValue(it.next()));
-                        }
-                        value = list;
-                    } else {
-                        Map<Object, Object> copyMap = new HashMap<Object, Object>((Integer)METHOD_SCRIPT_OBJECT_SIZE.invoke(value));
-                        for (Map.Entry<Object,Object> entry : (Iterable<Map.Entry<Object,Object>>) METHOD_SCRIPT_OBJECT_ENTRY_SET.invoke(value)) {
-                            copyMap.put(entry.getKey(), unwrapValue(entry.getValue()));
-                        }
-                        value = copyMap;
-                    }
-                } catch(InvocationTargetException ite) {
-                    throw new RuntimeException(ite);
-                } catch(IllegalAccessException iae) {
-                    throw new RuntimeException(iae);
-                }
-            } else if (value instanceof Map) {
-                Map<Object, Object> map = (Map<Object, Object>) value;
-                Map<Object, Object> copyMap = new HashMap<Object, Object>(map.size());
-                for (Map.Entry<Object,Object> entry : map.entrySet()) {
-                    copyMap.put(entry.getKey(), unwrapValue(entry.getValue()));
-                }
-                value = copyMap;
-            }
-            return value;
-        }
-
-        public static Object wrapValue(Object value) {
-            if (value == null) {
-                return null;
-            } else if (value instanceof Collection) {
-                Collection<Object> collection = (Collection<Object>) value;
-                Object[] array = new Object[collection.size()];
-                int index = 0;
-                for (Object obj : collection) {
-                    array[index++] = wrapValue(obj);
-                }
-                value = array;
-            }
-            return value;
-        }
-    }
-
-    private ScriptEngine nashorn;
-    private Bindings defaultContextBindings;
-    
     @Inject
     public JavascriptNashornScriptEngineService(Settings settings) {
         super(settings);
-        // setup the engine to share the definition of the Ecma script built-in objects: aka NashornGlobal.
-//        System.setProperty("nashorn.args", "--global-per-engine");
-//        ScriptEngineManager m = new ScriptEngineManager();
-//        nashorn = m.getEngineByName("nashorn");
-        // changing the system properties is not allowed by the tests. :(
-
         NashornScriptValueConverter.setup();
         
-//        NashornScriptEngineFactory factory = new NashornScriptEngineFactory();
-//        ScriptEngine engine = factory.getScriptEngine(new String[] { "--global-per-engine" });
-//        We cant use nashorn APIs so we need to go through introspection
+    }
+
+    /**
+     * @return The ScriptEngine to use. It is tied to the current thread and configured to reuse the same Global
+     * for all script executions on that thread.
+     */
+    private ScriptEngine getEngine() {
+        ScriptEngine engine = currentEngine.get();
+        if (engine != null) {
+        	return engine;
+        }
+//        System.err.println("Creating a new engine " + engineNumber++);
+        // setup the engine to share the definition of the Ecma script built-in objects: aka NashornGlobal.
+//      System.setProperty("nashorn.args", "--global-per-engine");
+//      ScriptEngineManager m = new ScriptEngineManager();
+//      nashorn = m.getEngineByName("nashorn");
+      // changing the system properties is not allowed by the tests. :(
+
+//      NashornScriptEngineFactory factory = new NashornScriptEngineFactory();
+//      ScriptEngine engine = factory.getScriptEngine(new String[] { "--global-per-engine" });
+//      We cant use nashorn APIs so we need to go through introspection
         try {
-        	String className = "jdk.nashorn.api.scripting.NashornScriptEngineFactory";
-        	Class<?> factoryClass = JavascriptPlugin.class.getClassLoader().loadClass(className);
-        	Method getScriptEngineMethod = factoryClass.getMethod("getScriptEngine", String[].class);
-        	nashorn = (ScriptEngine)getScriptEngineMethod.invoke(factoryClass.newInstance(), new Object[]{new String[] { "--global-per-engine" }});
+      	    String className = "jdk.nashorn.api.scripting.NashornScriptEngineFactory";
+	      	Class<?> factoryClass = this.getClass().getClassLoader().loadClass(className);
+	      	Method getScriptEngineMethod = factoryClass.getMethod("getScriptEngine", String[].class);
+	      	engine = (ScriptEngine)getScriptEngineMethod.invoke(factoryClass.newInstance(), new Object[]{new String[] { "--global-per-engine" }});
+	      	currentEngine.set(engine);
         } catch(Throwable t) {
         	t.printStackTrace();
         	throw new RuntimeException("Could not find the nashorn engine", t);
         }
-        defaultContextBindings = nashorn.getContext().getBindings(ScriptContext.ENGINE_SCOPE);
+        return engine;
     }
 
+    /**
+     * @return The CompiledScript to use. It is tied to the current thread.
+     */
+    private CompiledScript getCompiledScript(String script) {
+    	Map<String,CompiledScript> compiled = compiledScripts.get();
+    	if (compiled == null) {
+    		compiled = new HashMap<String, CompiledScript>();
+    		compiledScripts.set(compiled);
+    	}
+    	CompiledScript cs = compiled.get(script);
+    	if (cs != null) {
+    		return cs;
+    	}
+        try {
+        	Compilable compilableObj = (Compilable) getEngine();
+            cs = compilableObj.compile(script);
+            compiled.put(script, cs);
+            return cs;
+        } catch (ScriptException e) {
+            e.printStackTrace();
+            return null;
+        }
+    }
 
+    /**
+     * A Nashorn CompiledScript is only useful for a given Global object itself attached to the current Thread.
+	 * So we compile once in order to check that the script is well written.
+	 * But we don't cache it.
+     */
     @Override
     public Object compile(String script) {
         try {
-            Compilable compilableObj = (Compilable) nashorn;
-            return compilableObj.compile(script);
+        	Compilable compilableObj = (Compilable) getEngine();
+            compilableObj.compile(script);
+            return script;
         } catch (ScriptException e) {
             e.printStackTrace();
             return null;
@@ -177,21 +144,21 @@ public class JavascriptNashornScriptEngineService extends AbstractComponent
 
     @Override
     public ExecutableScript executable(Object compiledScript, Map<String, Object> vars) {
-        ScriptContext context = createContext(null, vars);
-        return new NashornExecutableScript((CompiledScript) compiledScript, context);
+        Bindings bindings = createBindings(null, vars);
+        ScriptContext context = createContext(bindings);
+        return new NashornExecutableScript(getCompiledScript((String)compiledScript), context, bindings);
     }
 
-    private ScriptContext createContext(SearchLookup lookup, Map<String, Object> vars) {
-        Bindings bindings = createBindings(lookup, vars);
-
-        ScriptContext newContext = new SilentScriptContext();
-        newContext.setBindings(defaultContextBindings, ScriptContext.ENGINE_SCOPE);
+    private ScriptContext createContext(Bindings bindings) {
+        ScriptContext newContext = new SimpleScriptContext();
+        Bindings builtinBindings = getEngine().getContext().getBindings(ScriptContext.ENGINE_SCOPE);
+        newContext.setBindings(builtinBindings, ScriptContext.ENGINE_SCOPE);
         newContext.setBindings(bindings, ScriptContext.GLOBAL_SCOPE);
         return newContext;
     }
     
     private Bindings createBindings(SearchLookup lookup, Map<String, Object> vars) {
-        Bindings bindings = nashorn.createBindings();
+        Bindings bindings = getEngine().createBindings();
         if (lookup != null) {
             for (Map.Entry<String, Object> entry : lookup.asMap().entrySet()) {
                 bindings.put(entry.getKey(), NashornScriptValueConverter.wrapValue(entry.getValue()));
@@ -208,8 +175,9 @@ public class JavascriptNashornScriptEngineService extends AbstractComponent
     @Override
     public Object execute(Object compiledScript, Map<String, Object> vars) {
         try {
-            ScriptContext context = createContext(null, vars);
-            return NashornScriptValueConverter.unwrapValue(((CompiledScript) compiledScript).eval(context));
+        	Bindings bindings = createBindings(null, vars);
+            ScriptContext context = createContext(bindings);
+            return NashornScriptValueConverter.unwrapValue((getCompiledScript((String)compiledScript)).eval(context));
         } catch (ScriptException e) {
             e.printStackTrace();
             return null;
@@ -239,90 +207,8 @@ public class JavascriptNashornScriptEngineService extends AbstractComponent
     @Override
     public SearchScript search(Object compiledScript, SearchLookup lookup,
             @Nullable Map<String, Object> vars) {
-        ScriptContext context = createContext(lookup, vars);
-        return new NashornSearchScript((CompiledScript) compiledScript, context, lookup);
-    }
-
-    public static class NashornExecutableScript implements ExecutableScript {
-
-        private final CompiledScript script;
-        private final ScriptContext context;
-        final Bindings bindings;
-        
-        public NashornExecutableScript(CompiledScript script, ScriptContext context) {
-            this.script = script;
-            this.context = context;
-            this.bindings = context.getBindings(ScriptContext.GLOBAL_SCOPE);
-        }
-
-        @Override
-        public Object run() {
-            try {
-                return unwrap(script.eval(context));
-            } catch (ScriptException e) {
-                e.printStackTrace();
-                return null;
-            }
-        }
-
-        @Override
-        public void setNextVar(String name, Object value) {
-            this.bindings.put(name, value);
-        }
-
-        @Override
-        public Object unwrap(Object value) {
-            return NashornScriptValueConverter.unwrapValue(value);
-        }
-    }
-
-    public static class NashornSearchScript extends NashornExecutableScript implements SearchScript {
-
-        private final SearchLookup lookup;
-        
-        public NashornSearchScript(CompiledScript script, ScriptContext context, SearchLookup lookup) {
-            super(script, context);
-            this.lookup = lookup;
-        }
-        
-        @Override
-        public void setNextReader(AtomicReaderContext reader) {
-            lookup.setNextReader(reader);
-        }
-
-        @Override
-        public void setScorer(Scorer scorer) {
-            lookup.setScorer(scorer);
-        }
-
-        @Override
-        public void setNextDocId(int doc) {
-            lookup.setNextDocId(doc);
-        }
-
-        @Override
-        public void setNextSource(Map<String, Object> source) {
-            lookup.source().setNextSource(source);
-        }
-
-        @Override
-        public void setNextScore(float score) {
-            this.bindings.put("_score", score);
-        }
-
-        @Override
-        public float runAsFloat() {
-            return ((Number) run()).floatValue();
-        }
-
-        @Override
-        public long runAsLong() {
-            return ((Number) run()).longValue();
-        }
-
-        @Override
-        public double runAsDouble() {
-            return ((Number) run()).doubleValue();
-        }
+    	Bindings bindings = createBindings(lookup, vars);
+        ScriptContext context = createContext(bindings);
+        return new NashornSearchScript(getCompiledScript((String)compiledScript), context, bindings, lookup);
     }
 }
